pyproject.toml
[project]
name = "pickup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "celery>=5.6.0",
    "cryptography>=45.0.5",
    "fastapi>=0.116.1",
    "gevent>=25.9.1",
    "greenlet>=3.2.3",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "msgspec>=0.20.0",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.10.1",
    "pydantic[email]==2.12.5",
    "pytest>=9.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-jose>=3.5.0",
    "python-redis-lock>=4.0.0",
    "python-socketio>=5.15.1",
    "python-ulid>=3.0.0",
    "redis>=6.2.0",
    "rich>=14.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
    "tzdata>=2025.2 ; sys_platform == 'win32'",
    "uvicorn>=0.35.0",
]


[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests/*"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["E226", "E401", "W292", "E712", "F405", "F403", "F401", "E731"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"  # use double quotes for strings.
line-ending = "auto"    # automatically detect the appropriate line ending.



# [[tool.uv.index]]
# name = "tuna"
# url = "https://pypi.tuna.tsinghua.edu.cn/simple"
# default = true   # 设为默认主索引源

